Don't overwrite manual post-processing delete folder preference

+++ medusa/process_tv.py
@@ -448,7 +448,7 @@ class ProcessResult(object):
             if self.unwanted_files:
                 self.delete_files(path, self.unwanted_files)
 
-            if not app.NO_DELETE or clean_folder:
+            if (proc_type != 'manual' and not app.NO_DELETE) or clean_folder:
                 check_empty = False if clean_folder else True
                 if self.delete_folder(path, check_empty=check_empty):
                     self.log_and_output('Deleted folder: {path}', level=logging.DEBUG, **{'path': path})
